+++ src/commands/ckbtc.rs
@@ -7,11 +7,7 @@ use ic_icrc1::Account;
 use openssl::sha::Sha256;
 use rust_decimal::Decimal;
 
-use crate::{
-    get_auth,
-    lib::{ckbtc_minter_canister_id, AnyhowResult},
-    BaseOpts,
-};
+use crate::lib::{ckbtc_minter_canister_id, AnyhowResult, AuthInfo};
 
 use super::print_vec;
 
@@ -25,44 +21,41 @@ mod withdrawal_address;
 /// Commands for chain-key bitcoin (ckBTC)
 #[derive(Subcommand)]
 pub enum CkbtcCommand {
-    Balance(BaseOpts<balance::BalanceOpts>),
-    UpdateBalance(BaseOpts<update_balance::UpdateBalanceOpts>),
-    Transfer(BaseOpts<transfer::TransferOpts>),
-    RetrieveBtc(BaseOpts<retrieve_btc::RetrieveBtcOpts>),
-    RetrieveBtcStatus(BaseOpts<retrieve_btc_status::RetrieveBtcStatusOpts>),
-    WithdrawalAddress(BaseOpts<withdrawal_address::GetWithdrawalAddressOpts>),
+    Balance(balance::BalanceOpts),
+    UpdateBalance(update_balance::UpdateBalanceOpts),
+    Transfer(transfer::TransferOpts),
+    RetrieveBtc(retrieve_btc::RetrieveBtcOpts),
+    RetrieveBtcStatus(retrieve_btc_status::RetrieveBtcStatusOpts),
+    WithdrawalAddress(withdrawal_address::GetWithdrawalAddressOpts),
 }
 
-pub fn dispatch(command: CkbtcCommand) -> AnyhowResult {
+pub fn dispatch(
+    auth: &AuthInfo,
+    command: CkbtcCommand,
+    qr: bool,
+    fetch_root_key: bool,
+) -> AnyhowResult {
     match command {
         CkbtcCommand::UpdateBalance(opts) => {
-            let qr = opts.global_opts.qr;
-            let out = update_balance::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            let out = update_balance::exec(auth, opts)?;
             print_vec(qr, &out)?;
         }
         CkbtcCommand::Transfer(opts) => {
-            let qr = opts.global_opts.qr;
-            let out = transfer::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            let out = transfer::exec(auth, opts)?;
             print_vec(qr, &out)?;
         }
         CkbtcCommand::RetrieveBtc(opts) => {
-            let qr = opts.global_opts.qr;
-            let out = retrieve_btc::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            let out = retrieve_btc::exec(auth, opts)?;
             print_vec(qr, &out)?;
         }
         CkbtcCommand::RetrieveBtcStatus(opts) => {
-            retrieve_btc_status::exec(opts.command_opts, opts.global_opts.fetch_root_key)?;
+            retrieve_btc_status::exec(opts, fetch_root_key)?;
         }
         CkbtcCommand::Balance(opts) => {
-            let fetch_root_key = opts.global_opts.fetch_root_key;
-            balance::exec(
-                &get_auth(opts.global_opts)?,
-                opts.command_opts,
-                fetch_root_key,
-            )?;
+            balance::exec(auth, opts, fetch_root_key)?;
         }
         CkbtcCommand::WithdrawalAddress(opts) => {
-            withdrawal_address::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            withdrawal_address::exec(auth, opts)?;
         }
     }
     Ok(())