+++ tests/commands/sns-neuron-stake-memo.sh
@@ -1,0 +1,1 @@
+"$QUILL" sns stake-neuron --amount 12 --memo 777 --canister-ids-file ./sns_canister_ids.json --pem-file - | "$QUILL" send --dry-run -