+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - ckBTC commands and support. (#153)
+- SNS commands and support (replaces sns-quill). (#154)
 
 ## [0.3.2] - 2023-01-13